Continued work on the start script
