+++ packages/api-pls/bin/commands/start.js
@@ -3,10 +3,9 @@
 const path = require('path');
 const chalk = require('chalk');
 
-module.exports = function() {
+module.exports = function(options) {
   console.log(chalk.grey('Starting the API webserver...'));
-
   const serverPath = path.join(__dirname, '..', '..', 'index.js');
-
-  require(serverPath);
+  const server = require(serverPath);
+  server(options);
 };
