+++ packages/api-pls/bin/pls.js
@@ -7,6 +7,11 @@ const migrate = require('./commands/migrate');
 const resetDatabase = require('./commands/reset-database');
 const start = require('./commands/start');
 
+// TODO: Remove this usage of dotenv. Currently, this is used for one thing:
+// getting the database URL. This needs to passed in through some other means.
+const envPath = global.ENV_PATH ? global.ENV_PATH : '.env';
+require('dotenv').config({path: envPath});
+
 // Load up our configuration, passing the defaults in.
 let options = rc('pls', {
   resourcesDirectory: "./resources"
