+++ packages/api-pls/package.json
@@ -54,7 +54,6 @@
     "helmet": "3.4.0",
     "inquirer": "^3.0.1",
     "is-my-json-valid": "2.15.0",
-    "js-yaml": "3.7.0",
     "kapow": "1.0.0",
     "lodash": "4.17.4",
     "path": "0.12.7",
