--- packages/api-pls/server/util/load-resource-configs.js
@@ -1,14 +1,0 @@
-'use strict';
-
-module.exports = function() {
-  const resourceDefinitions = process.argv[2];
-
-  let resources = [];
-  try {
-    resources = JSON.parse(resourceDefinitions);
-  } catch (e) {
-    // Intentionally blank
-  }
-
-  return resources;
-};
